kubernetes-auth-server/requirements.txt
flask
python-dotenv
requests
flask-oauthlib
python-jose